APIGOV-21514 - move api exception list to its own file
