+++ pkg/traceability/config.go
@@ -2,8 +2,6 @@ package traceability
 
 import (
 	"net/url"
-	"regexp"
-	"strings"
 	"time"
 
 	"github.com/Axway/agent-sdk/pkg/agent"
@@ -119,28 +117,10 @@ func readConfig(cfg *common.Config, info beat.Info) (*Config, error) {
 		}
 	}
 
-	// if set check for valid regex definitions
-	if outputConfig.APIExceptionsList != nil {
-		newShowRegex = make([]showRegex, 0)
-
-		// Get the api exceptions list
-		exceptions := outputConfig.APIExceptionsList
-		for i := range exceptions {
-			exception := strings.TrimSpace(exceptions[i])
-
-			// check for regex and then validate
-			keyMatch, err := regexp.Compile(exception)
-			if err != nil {
-				err = ErrInvalidRegex.FormatError("apiExceptionValue", exception, err)
-				log.Error(err)
-				return nil, err
-			}
-
-			newShowRegex = append(newShowRegex, showRegex{
-				keyMatch: keyMatch,
-			})
-
-		}
+	exception, err := setUpAPIExceptionList(outputConfig.APIExceptionsList)
+	if err != nil {
+		err = ErrInvalidRegex.FormatError("apiExceptionValue", exception, err)
+		log.Error(err)
 	}
 
 	return outputConfig, nil
@@ -160,20 +140,4 @@ func GetMaxRetries() int {
 		return 3
 	}
 	return outputConfig.MaxRetries
-}
-
-// getAPIExceptionsList - Returns traceability APIs exception list (api paths)
-func getAPIExceptionsList() []showRegex {
-	if outputConfig == nil {
-		return []showRegex{}
-	}
-	return newShowRegex
-}
-
-// newShowRegex - array of showRegex
-var newShowRegex []showRegex
-
-// showRegex - regular expression to check against
-type showRegex struct {
-	keyMatch *regexp.Regexp
 }
