+++ pkg/traceability/traceability.go
@@ -372,18 +372,3 @@ func (ta *traceabilityAgentHealthChecker) connectionHealthcheck(host string) *hc
 	}
 	return status
 }
-
-// ShouldIgnoreEvent - check to see if the uri exists in exception list
-func ShouldIgnoreEvent(uriRaw string) bool {
-	exceptions := getAPIExceptionsList()
-
-	// If the api path exists in the exceptions list, return true and ignore event
-	for _, exception := range exceptions {
-		if exception.keyMatch.MatchString(uriRaw) {
-			return true
-		}
-	}
-
-	// api path not found in exceptions list
-	return false
-}
